Move all assert helps into support.asserts.
For a better management of assertion helpers those should be
added to the central assertion helper module.

MozReview-Commit-ID: DYMfOFl9Fjd

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1433390 [ci skip]
